refractor: add seperate parser
